+++ packages/bridge-controller/src/utils/fetch.ts
@@ -12,6 +12,7 @@ import type {
   QuoteStreamCompleteData,
   BatchSellTradesRequest,
   BatchSellTradesResponse,
+  FeatureId,
 } from '../types';
 import { getEthUsdtResetData } from './bridge';
 import {
@@ -21,7 +22,6 @@ import {
 } from './caip-formatters';
 import { fetchServerEvents } from './fetch-server-events';
 import { isEvmTxData } from './trade-utils';
-import type { FeatureId } from './validators';
 import {
   validateQuoteResponse,
   validateSwapsTokenObject,
@@ -325,6 +325,7 @@ const getQuoteResponseId = ({
  * @param fetchFn - The fetch function to use
  * @param quoteRequests - An array of GenericQuoteRequest objects
  * @param signal - The abort signal
+ * @param featureId - The {@link FeatureId} for the experience that's requesting the quotes
  * @param clientId - The client ID for metrics
  * @param jwt - The JWT token for authentication
  * @param bridgeApiBaseUrl - The base URL for the bridge API
@@ -341,6 +342,7 @@ export async function fetchBridgeQuoteStream(
   fetchFn: FetchFunction,
   quoteRequests: GenericQuoteRequest[],
   signal: AbortSignal | undefined,
+  featureId: FeatureId,
   clientId: string,
   jwt: string | undefined,
   bridgeApiBaseUrl: string,
@@ -380,6 +382,7 @@ export async function fetchBridgeQuoteStream(
 
         return await serverEventHandlers.onValidQuoteReceived({
           ...quoteResponse,
+          featureId,
           // Append the reset approval data to the quote response if the request has resetApproval set to true and the quote has an approval
           resetApproval:
             matchingQuoteRequest.resetApproval &&
@@ -495,6 +498,7 @@ export async function fetchBridgeQuoteStream(
 
 export const formatBatchSellTradesRequest = (
   quotes: (QuoteResponse | null)[],
+  stxEnabled: boolean,
 ): BatchSellTradesRequest => ({
   quotes: quotes
     .filter((quote): quote is QuoteResponse => quote !== null)
@@ -513,12 +517,14 @@ export const formatBatchSellTradesRequest = (
         quoteId,
       }),
     ),
+  stxEnabled,
 });
 
 /**
  * Fetches quotes from the bridge-api's getQuote endpoint
  *
  * @param quotes - The quotes to fetch the gasless transaction data and fees for. May contain null values if a quote is not available for a swap
+ * @param stxEnabled - Flag to estimate gas cost more precisely for the batch sell feature.
  * @param signal - The abort signal
  * @param clientId - The client ID for metrics
  * @param jwt - The JWT token for authentication
@@ -529,6 +535,7 @@ export const formatBatchSellTradesRequest = (
  */
 export async function fetchBatchSellTrades(
   quotes: (QuoteResponse | null)[],
+  stxEnabled: boolean,
   signal: AbortSignal | null,
   clientId: string,
   jwt: string | undefined,
@@ -537,7 +544,10 @@ export async function fetchBatchSellTrades(
   clientVersion?: string,
 ): Promise<BatchSellTradesResponse> {
   const url = `${bridgeApiBaseUrl}/obtainGaslessBatch`;
-  const request: BatchSellTradesRequest = formatBatchSellTradesRequest(quotes);
+  const request: BatchSellTradesRequest = formatBatchSellTradesRequest(
+    quotes,
+    stxEnabled,
+  );
   const batchSellTradesResponse = await fetchFn(url, {
     headers: {
       ...getClientHeaders({